[ refactor ] : 사용하지않는 클래스 제거
