+++ app/src/main/java/com/ddd/ansayo/common/CourseViewHolder.kt
@@ -1,8 +1,8 @@
 package com.ddd.ansayo.common
 
 import androidx.recyclerview.widget.RecyclerView
+import com.ddd.ansayo.core_model.course.Course
 import com.ddd.ansayo.databinding.ItemCourseBinding
-import com.ddd.ansayo.domain.model.course.Course
 
 class CourseViewHolder(
     private val binding: ItemCourseBinding,
