--- domain/src/main/java/com/ddd/ansayo/domain/model/course/Course.kt
@@ -1,5 +1,0 @@
-package com.ddd.ansayo.domain.model.course
-
-data class Course(
-    val todo: Int
-)
